requirements.txt
tensorflow
numpy
pillow
tkinter
scipy